backend/requirements.txt
# 文件作用：后端（FastAPI）所需的基础依赖清单，便于可重复安装与环境一致性。
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
loguru==0.7.2
python-dotenv==1.0.1
redis==5.0.7

# 说明：如需 SSE 扩展库或数据库驱动，请在后续阶段按需追加。
PyJWT==2.9.0
prometheus-client==0.21.0
fastapi-cors==0.0.6

